% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_bloomberg_file.R
\name{process_bloomberg_file}
\alias{process_bloomberg_file}
\title{Process Bloomberg File}
\usage{
process_bloomberg_file(year)
}
\arguments{
\item{year}{Numeric value of the year of the file you want to read in.}
}
\value{
A data frame with selected variables.
}
\description{
Process Bloomberg File
}
\examples{
x <- process_bloomberg_file(2005)
}

